# specs/001-inventory-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: POS Inventory Management API
  version: 1.0.0
  description: |
    HTTP contract for the POS Inventory subsystem. All endpoints require a
    bearer JWT minted by the surrounding POS auth service. The token MUST
    carry `tenant_id` and `roles[]` constrained to the canonical role set
    (Cashier, Receiver, Inventory Clerk, Store Manager, Purchasing, Admin).
servers:
  - url: https://{host}/v1
    variables:
      host: { default: api.example.com }
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UUID: { type: string, format: uuid }
...
    Quantity: { type: string, pattern: '^-?\d+(\.\d{1,3})?$' }
...
    VendorRmaInput:
      type: object
      required: [vendor_id, lines]
      properties:
        vendor_id: { $ref: '#/components/schemas/UUID' }
        originating_purchase_order_id: { $ref: '#/components/schemas/UUID', nullable: true }
        lines:
          type: array
          items:
            type: object
            required: [sku_id, qty]
            properties:
              sku_id: { $ref: '#/components/schemas/UUID' }
              qty: { $ref: '#/components/schemas/Quantity' }
              serial_value: { type: string, nullable: true }
              lot_code: { type: string, nullable: true }
              customer_return_line_id: { $ref: '#/components/schemas/UUID', nullable: true }
...
paths:
...
  /vendor-rmas:
    post:
      summary: Create a vendor RMA (Inventory Clerk or Store Manager) — FR-017.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VendorRmaInput' } } }
      responses:
        '201': { description: Created }
  /vendor-rmas/{id}/ship:
    parameters: [ { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } } ]
    post:
      summary: open → shipped (Inventory Clerk or Store Manager) — FR-017, FR-018.
      responses: { '200': { description: OK } }
  /vendor-rmas/{id}/close:
    parameters: [ { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } } ]
    post:
      summary: shipped → closed; serials advance to rma_closed (FR-018).
      responses: { '200': { description: OK } }
...